requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
faiss-cpu
sentence-transformers
numpy
python-dotenv
redis
requests
openai
anthropic
spacy